+++ src/lib/pages/contract-details/index.tsx
@@ -114,10 +114,20 @@ const ContractDetails = () => {
         <InstantiateInfo contractDetail={contractDetails} />
         {/* Contract Info (Expand) */}
         <Flex direction="column" flex={0.8} gap={4}>
-          <JsonInfo header="Contract Info" jsonString="" />
+          <JsonInfo
+            header="Contract Info"
+            jsonString={JSON.stringify(
+              contractDetails?.instantiateInfo?.raw.contract_info ?? "",
+              null,
+              2
+            )}
+            jsonAreaHeight="180px"
+          />
           <JsonInfo
             header="Instantiate Messages"
             jsonString={jsonPrettify(contractDetails?.initMsg ?? "")}
+            showViewFullButton
+            defaultExpand
           />
         </Flex>
       </Flex>